+++ miden-crypto/src/merkle/smt/forest/store.rs
@@ -169,9 +169,9 @@ impl SmtStore {
         #[allow(unused_mut)]
         let mut sorted_leaf_indices = leaves_by_index.keys().cloned().collect::<Vec<_>>();
 
-        #[cfg(feature = "hashmaps")]
+        #[cfg(feature = "std")]
         // Sort leaves by NodeIndex to easily detect when leaves share a parent (only neighboring
-        // leaves can share a parent). Hashbrown::HashMap doesn't maintain key ordering, so
+        // leaves can share a parent). std::collections::HashMap doesn't maintain key ordering, so
         // we need to sort the indices.
         sorted_leaf_indices.sort();